+++ Data/src/utility/TemperatureSensor.h
@@ -3,18 +3,19 @@
 
 #include <Arduino.h>
 #include "Sensor.h"
+
 #define MODE_C 0
 #define MODE_F 1
 
 class TemperatureSensor : public Sensor {
 
   public:
-    TemperatureSensor(String name, int pin, int mode = MODE_C);
+    TemperatureSensor(String name, byte pin, int mode = MODE_C);
     float readValue();
     int setMode(int mode);
 	
   private:
-	int _pin;
+	byte _pin;
 	int _mode;
 };
